Moves batches into tasks dir
