+++ lib/tasks/digest.js
@@ -7,7 +7,7 @@ const fetchData = async () => {
 
   const res = await airtable.get(
     tableName,
-    { filterByFormula: 'IS_AFTER(CREATED_TIME(),(DATEADD(TODAY(), -1, "hours")))'}
+    { filterByFormula: 'IS_AFTER(LAST_MODIFIED_TIME(),(DATEADD(TODAY(), -1, "hours")))'}
   );
 
   const { records } = res.body;
