+++ worker.js
@@ -5,7 +5,7 @@ require('dotenv').config();
 const { argv } = require('yargs')
 const logger = require('heroku-logger');
 
-const batches = require('./lib/batches');
+const batches = require('./lib/tasks/batches');
 const digest = require('./lib/tasks/digest');
 
 // Places new subscribers into a batch group, if they haven't been added to one yet.
@@ -13,6 +13,8 @@ const main = async () => {
   try {
     if (argv.task === 'digest') {
       const result = await digest.send();
+
+      logger.debug('Digest', { result });
 
       return result;
     }
